refactor: Add annotations, move switch to toggles.py
