+++ openedx/core/djangoapps/oauth_dispatch/views.py
@@ -15,9 +15,10 @@ from ratelimit import ALL
 from ratelimit.decorators import ratelimit
 
 from openedx.core.djangoapps.auth_exchange import views as auth_exchange_views
-from openedx.core.djangoapps.oauth_dispatch import adapters, DISABLE_JWT_FOR_MOBILE
+from openedx.core.djangoapps.oauth_dispatch import adapters
 from openedx.core.djangoapps.oauth_dispatch.dot_overrides import views as dot_overrides_views
 from openedx.core.djangoapps.oauth_dispatch.jwt import create_jwt_token_dict
+from openedx.core.djangoapps.oauth_dispatch.toggles import DISABLE_JWT_FOR_MOBILE
 from openedx.core.lib.mobile_utils import is_request_from_mobile_app
